+++ taotao-manager-mapper/src/main/java/com/taotao/mapper/TbRoleMenuMapper.java
@@ -10,7 +10,7 @@ public interface TbRoleMenuMapper {
 
     int deleteByExample(TbRoleMenuExample example);
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(TbRoleMenu record);
 
@@ -18,7 +18,7 @@ public interface TbRoleMenuMapper {
 
     List<TbRoleMenu> selectByExample(TbRoleMenuExample example);
 
-    TbRoleMenu selectByPrimaryKey(Long id);
+    TbRoleMenu selectByPrimaryKey(Integer id);
 
     int updateByExampleSelective(@Param("record") TbRoleMenu record, @Param("example") TbRoleMenuExample example);
